Introduce a mechanism that allows each prop to have a permalink via HTML anchors.
Under the hood the rendering components get access to a 'componentDisplayName' which can be used as a composed 'id' to create unique anchor values.
For all out-of-the-box layouts this has been implemented as a simple change of background-color, for a custom 'LayoutRenderer' you can just use whatever styles you prefer.